+++ CHANGELOG.md
@@ -9,7 +9,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 
-## [1.1.0] - 2024-10-11
+## [2.0.1] - 2024-10-15
+
+### Fixed
+
+- **Database Connection Issues**:
+  Fixed issue with users connecting to LabMateAI database.
+
+## [2.0.0] - 2024-10-11
 
 ### Added